web sistem/requirements.txt
flask
opencv-python
joblib
pillow
scikit-learn
scikit-image
pandas